feat: Moved Navbar component to layout.tsx file for consistent visual and made it responsive to smaller screens
